backend/requirements.txt
Flask
Werkzeug
Flask-CORS
google-auth-oauthlib
pymongo
requests
bcrypt
pyjwt
